pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "wachitugo",email = "benjaruiz7028@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"

dependencies = [
    "pandas (>=2.3.0,<3.0.0)",
    "langgraph (>=0.5.0,<0.6.0)",
    "fpdf (>=1.7.2,<2.0.0)",
    "nbformat (>=5.10.4,<6.0.0)",
    "langchain-openai (>=0.3.27,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "ipywidgets (>=8.1.7,<9.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-community (>=0.3.26,<0.4.0)"
    
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
